+++ packages/react/src/components/DataTable/src/components/DataTableCell/DataTableCell.tsx
@@ -7,7 +7,8 @@ import { DataTableCellProps } from "./dataTableCell.types";
 const DataTableCell: React.FC<DataTableCellProps> = ({
   className: _className,
   style: _style,
+  children,
   ...rest
-}: DataTableCellProps) => <Table.Cell {...rest} />;
+}: DataTableCellProps) => <Table.Cell {...rest}>{children}</Table.Cell>;
 
 export { DataTableCell };