+++ packages/react/src/components/DataTable/src/components/DataTableCell/dataTableCell.types.ts
@@ -1,6 +1,15 @@
-import { HTMLAttributes } from "react";
+import { HTMLAttributes, ReactNode } from "react";
 
 import { TableCellProperties } from "@nimbus-ds/components";
 
-export type DataTableCellProps = Omit<HTMLAttributes<HTMLElement>, "color"> &
+export interface DataTableCellProperties {
+  /**
+   * Content of the List component.
+   * @TJS-type React.ReactNode
+   */
+  children: ReactNode;
+}
+
+export type DataTableCellProps = DataTableCellProperties &
+  Omit<HTMLAttributes<HTMLElement>, "color"> &
   TableCellProperties;